Move designator and literal tests to expressions test file
